Introduce log system (and more).
First, Generic controller extends Hoathis\Kit\Aggregator. Then, we no
longer use $_this but directly and simply $this. It saves also memory
and CPU usage (because we do not instantiate two kits, and because we
avoid reflection to distribute $_this on action methods).

Second, we create Log controller, with its associated views.

Finally, we add session control. Use $this->adminGuard() in controller
(located in the Generic controller) to protect a page.